+++ src/indexing/linkTitle.py
@@ -8,7 +8,7 @@ import whoosh.index
 from . import getIndexPath
 
 
-def index(corpusPath, name, tweetTime = None, stored = False):
+def index(corpusPath, name, tweetTime = None, stored = False, overwrite = True):
     """Indexing of titles of the linked pages."""
     
     dirList = os.listdir(corpusPath)
@@ -22,6 +22,8 @@ def index(corpusPath, name, tweetTime = None, stored = False):
     if not os.path.exists(indexPath):
         os.makedirs(indexPath)
     else:
+        if not overwrite:
+            return
         shutil.rmtree(indexPath)
         os.makedirs(indexPath)
     ix = whoosh.index.create_in(indexPath, schema)
